OkHttp: Drop leak warning message that may not be the app's fault.
When a HttpURLConnection is disconnect()ed but its response body
isn't closed, OkHttp logs a warning because it delays garbage
collection of associated state. One way this can happen is if an
app only accesses the response header fields; this is not an
app problem so we shouldn't warn about it.

Bug: 64789755
Test: Treehugger

Change-Id: If809b498d5567b68968eee6bf1be5e759eca677c