Change OkHttp test infrastructure to not use BC provider
The OkHttp test infrastructure specifically requests the implementations from
the BC provider, most likely because they wanted their tests to behave the
same on desktops as on devices, and Bouncy Castle used to be the default
provider of most or all required functions.  With BC's implementations being
deprecated and most users relying on the Conscrypt implementations instead
(since they're the default), switch to the Conscrypt implementations.

Bug: 67626877
Test: cts -m CtsLibcoreOkHttpTestCases
Change-Id: I83b26ed5aa9bab34eedaf6690cbfe40101b360d8