Move of OkHttp facade API to new package.
This CL moves the classes that make up this facade from
  libcore.net.http
to
  com.squareup.okhttp.internalandroidapi.
In jarjar'ed build targets, this becomes
  com.android.okhttp.internalandroidapi.

This facade constitutes the API via which non-libcore parts
of the Android platform (currently framework) may access
OkHttp. It's moving because libcore.net.http is already
part of libcore, and the overlap of packages is problematic for
builds with EXPERIMENTAL_USE_OPENJDK9 set to true.

Http{,s}Handler are not currently imported by framework code;
this CL moves them back to com.android.okhttp. That package is
also where preloaded-classes still expected them to be.

Bug: 68220880
Test: CtsLibcoreOkHttpTestCases
Change-Id: I6f01c1d0f59db43290dac126f6b6ed2071d7708a