Annotate repackaged API with @UnsupportedAppUsage
Adds @UnsupportedAppUsage annotations to the parts of okhttp that
are in the hidden API. Allows the okhttp entries to be removed from
the frameworks/base/config/hiddenapi-light-greylist.txt.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Test: see above
Change-Id: I5c88c7a80ee1d2ec7b688ae8175267fac35ef368