Relax test constraint for disconnected HttpURLConnection.
During test the MockWebServer when exactly half of the request
has been read, and the test then asserts that the amouint of data
written to the socket before receiving an exception is within 1%
of the total.  Automated testing on Taimen (see bug) sometimes
manages to write more which makes the test flake. As there is
always going to some non-determinism in IO buffering, bumping the
tolerance from 1% to 2% seems reasonable.

Bug: 126877539
Test: atest CtsLibcoreOkHttpTestCases
Change-Id: I1e11a5f31d499334e50d80505ebc0747c2a56d90