Set default visibility to private
Takes advantage of a new mechanism for specifying the default visibility
of all modules in a package to restrict the visibility of okhttp modules
as much as possible.

Added visibility to okhttp-testdex even though it is only used in .mk
files to record the dependency as the name is constructed dynamically by
adding "-testdex" suffix to a list of other dependencies and so cannot
be found by searching.

Bug: 134379140
Test: m droid
Change-Id: I5cd5487c57c4ca6580cc052feb24e12f20e02b89